move tests for fakeServer into existing test file
