+++ lib/sinon/util/fake_server.js
@@ -23,12 +23,6 @@ if (typeof sinon == "undefined") {
 
 (function () {
     var push = [].push;
-    function F() {}
-
-    function create(proto) {
-        F.prototype = proto;
-        return new F();
-    }
 
     function responseArray(handler) {
         var response = handler;
@@ -80,7 +74,7 @@ if (typeof sinon == "undefined") {
     function makeApi(sinon) {
         sinon.fakeServer = {
             create: function (config) {
-                var server = create(this);
+                var server = sinon.create(this);
                 server.configure(config);
                 if (!sinon.xhr.supportsCORS) {
                     this.xhr = sinon.useFakeXDomainRequest();
